Bioinformatics Practice 1
Task: Given a SAM file and a reference fasta file, the goal is to implement a function that will create a pileup of mapped reads over a reference sequence. A function will also be written to identify potential sequencing errors in reads of equal length with the same mapping position.